Squashed 'lemon/' changes from bfbc1dc..0daaf4b
0daaf4b fix bugs
117eacd Merge branch 'develop' of https://github.com/gsmake/gsmake into develop
f5df3fc fix some bugs
45d0d62 change gc algorithm
b996798 add ref counter gc
9fb0575 implement reactor_io_socket
78c96d5 daily commit
30a2ee4 update
b440c70 update
ccf7b40 update
d3fa516 Merge commit 'b9f0df1cf441ef2e7b0dc3f3d56500956c1b402d' into develop
cd17939 fix posix copy_file file handler leak bug
d6d2504 implement posix fs:status function
0dcdc1e fix some bug
c7542e4 update
44dc59e fix some bugs
5f6ac90 fix some bugs

git-subtree-dir: lemon
git-subtree-split: 0daaf4bbec0e128756a1f7fbf10ceec6282bd826

+++ src/main/cpp/lemoon/fs.cpp
@@ -105,9 +105,9 @@ namespace lemoon{namespace{
 
         auto path = luaL_checkstring(L,1);
 
-        if(lua_type(L,2) != LUA_TNONE && lua_type(L,2) != LUA_TNIL)
+        if(lua_type(L,2) == LUA_TBOOLEAN && lua_toboolean(L,2) )
         {
-            if(lemon::fs::is_directory(path))
+            if(lemon::fs::is_directory(path) && !lemon::fs::is_symlink(path))
             {
                 lemon::fs::remove_directories(path,err);
 
@@ -124,7 +124,7 @@ namespace lemoon{namespace{
 
         if(err)
         {
-            return luaL_error(L,"[lemoon.fs] call emon::fs::remove_file error :%s",err.message().c_str());
+            return luaL_error(L,"[lemoon.fs] remove file(%s) error :%s",path,err.message().c_str());
         }
 
         return 0;
